Criando o arquivo conftest que possui as fixtures sessao e conexao
close #32
