+++ libpython/spam/modelos.py
@@ -1,0 +1,4 @@
+class Usuario:
+    def __init__(self, nome):
+        self.nome = nome
+        self.id = None
