+++ libpython/tests/test_spam/test_usuarios.py
@@ -1,29 +1,4 @@
-import pytest
-
-from libpython.spam.db import Conexao
-
-
-@pytest.fixture
-def conexao():
-    # Setup
-    conexao_obj = Conexao()
-    yield conexao_obj
-    # Tear Down
-    conexao_obj.fechar()
-
-
-@pytest.fixture
-def sessao(conexao):
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    sessao_obj.rollback()
-    sessao_obj.fechar()
-
-
-class Usuario:
-    def __init__(self, nome):
-        self.nome = nome
-        self.id = None
+from libpython.spam.modelos import Usuario
 
 
 def test_salvar_usuario(conexao, sessao):
